Add files via upload
A login and registration form in C++ using file handling can be created to take input arguments of username and password. The program will check if the user is already registered or not and create a new account if they are not. Usernames and passwords can be stored in a text file named "register.txt" in the format "username,password". The program will check if the entered password matches the one in the file and grant access if they match. This can be accomplished with simple code using file handling in C++